Fix messagekit
